% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahull.IUCN.R
\name{ahull.IUCN}
\alias{ahull.IUCN}
\title{Ahull according to the IUCN Red List Categories and Criteria}
\usage{
ahull.IUCN(x, y = NULL, alpha)
}
\arguments{
\item{x}{The x and y arguments provide the x and y coordinates of a set of points. Alternatively, a single argument x can be provided.}

\item{y}{The x and y arguments provide the x and y coordinates of a set of points. Alternatively, a single argument x can be provided.}

\item{alpha}{A number.}
}
\value{
An ahull.IUCN object.
}
\description{
Ahull according to the IUCN Red List Categories and Criteria
}
\examples{
n <- 500
theta<-runif(n,0,2*pi)
r<-sqrt(runif(n,0.25^2,0.5^2))
x<-cbind(0.5+r*cos(theta),0.5+r*sin(theta))
alpha<-2.5
ah.IUCN.obj<-ahull.IUCN(x,alpha=alpha)
}
\references{
IUCN. 2022. The IUCN Red List of Threatened Species. Version 2022-2. https://www.iucnredlist.org

Burgman, M., & Fox, J. (2003). Bias in species range estimates from minimum convex polygons: Implications for conservation and options for improved planning. Animal Conservation Forum, 6(1), 19-28. doi:10.1017/S1367943003003044
}
