\name{alphahull-package}
\alias{alphahull-package}
\docType{package}
\title{
Generalization of the convex hull of a sample of points in the plane
}
\description{
This package computes the \eqn{\alpha}-shape and \eqn{\alpha}-convex hull of a given sample of points in the plane. The concepts of \eqn{\alpha}-shape and \eqn{\alpha}-convex hull generalize the definition of the convex hull of a finite set of points. The programming is based on the duality between the Voronoi diagram and Delaunay triangulation. The package also includes a function that returns the Delaunay mesh of a given sample of points and its dual Voronoi diagram in one single object.}
\details{
\tabular{ll}{
Package: \tab alphahull\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2011-10-21\cr
License: \tab R functions: file LICENSE
\cr
}
}
\author{
Beatriz Pateiro-Lopez, Alberto Rodriguez-Casal.

Maintainer: Beatriz Pateiro-Lopez <beatriz.pateiro@usc.es>
}

\keyword{nonparametric}