% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaahull.R
\name{areaahull}
\alias{areaahull}
\title{Area of the alpha-convex hull}
\usage{
areaahull(x, timeout = 5)
}
\arguments{
\item{x}{Object of class \code{"ahull"}.}

\item{timeout}{A numeric specifying the maximum number of seconds the
expression is allowed to run before being interrupted by the timeout.}
}
\value{
\item{area}{Area of the \eqn{\alpha}-convex hull. If the area cannot
be computed, the output will be NA with a warning.}
}
\description{
This function calculates the area of the \eqn{\alpha}-convex hull of a
sample of points.
}
\examples{

\dontrun{
# Random sample in the unit square
x <- matrix(runif(500), nc = 2)
# Value of alpha
alpha <- 1
# alpha-convex hull
ahull.obj <- ahull(x, alpha = alpha)
# Area of the alpha-convex hull
areaahull(ahull.obj)
}

}
\seealso{
\code{\link{ahull}}.
}
\keyword{nonparametric}
