% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaahulleval.R
\name{areaahulleval}
\alias{areaahulleval}
\title{Area of the alpha-convex hull}
\usage{
areaahulleval(x)
}
\arguments{
\item{x}{Object of class \code{"ahull"}.}
}
\value{
\item{area}{Area of the \eqn{\alpha}-convex hull. }
}
\description{
This function calculates the area of the \eqn{\alpha}-convex hull of a
sample of points.
}
\seealso{
\code{\link{ahull}}.
}
\keyword{internal}
