% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthahull.R
\name{lengthahull}
\alias{lengthahull}
\title{Length of the boundary of the alpha-convex hull}
\usage{
lengthahull(ahull.arcs)
}
\arguments{
\item{ahull.arcs}{Output matrix of arcs returned by \code{\link{ahull}}.}
}
\value{
\item{length}{Length of the boundary of the \eqn{\alpha}-convex
hull.}
}
\description{
This function calculates the length of the boundary of the
\eqn{\alpha}-convex hull of a given sample.
}
\details{
The function \code{lengthahull} is internally called by the function
\code{\link{ahull}}.
}
\examples{

\dontrun{
# Random sample in the unit square
x <- matrix(runif(100), nc = 2)
# Value of alpha
alpha <- 0.2
# alpha-convex hull
ahull.obj <- ahull(x, alpha = alpha)
# Length of the alpha-convex hull
ahull.obj$length
}

}
\seealso{
\code{\link{ahull}}.
}
\keyword{nonparametric}
