% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ahull.IUCN.R
\name{plot.ahull.IUCN}
\alias{plot.ahull.IUCN}
\title{Very simple plot function}
\usage{
\method{plot}{ahull.IUCN}(x, wplot = c("all", "comp"), ...)
}
\arguments{
\item{x}{The ahull.IUCN object.}

\item{wplot}{"Which in the plot?". I.e. should all points segments and connected components be plat (wlines='all', the default), should only connected components be plotted (wlines='comp')}

\item{...}{Additional plot parameters.}
}
\value{
None.
}
\description{
Very simple plot function
}
\examples{
n <- 500
theta<-runif(n,0,2*pi)
r<-sqrt(runif(n,0.25^2,0.5^2))
x<-cbind(0.5+r*cos(theta),0.5+r*sin(theta))
alpha<-2.5
ah<-ahull.IUCN(x,alpha=alpha)
plot(ah)
}
