% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkoch.R
\name{rkoch}
\alias{rkoch}
\title{Random generation on a Koch snowflake curve}
\usage{
rkoch(n, side = 3, niter = 5)
}
\arguments{
\item{n}{Number of observations.}

\item{side}{Side length of the initial equilateral triangle of the Koch
snowflake curve.}

\item{niter}{Number of iterations in the development of the snowflake curve.}
}
\value{
A 2-column matrix with the coordinates of generated points.
}
\description{
This function generates ramdom points from a uniform distribution on a Koch
snowflake.
}
\examples{

\dontrun{
unifkoch <- rkoch(2000, side = 1, niter = 3)
plot(unifkoch, asp = TRUE)
}

}
\seealso{
\code{\link{koch}}.
}
\keyword{nonparametric}
