% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\title{Clockwise rotation}
\usage{
rotation(v, theta)
}
\arguments{
\item{v}{Vector \eqn{v} in the plane.}

\item{theta}{Angle \eqn{\theta} (in radians).}
}
\value{
\item{v.rot}{Vector after rotation.}
}
\description{
This function calculates the clockwise rotation of angle \eqn{\theta} of a
given vector \eqn{v} in the plane.
}
\examples{

\dontrun{
# Rotation of angle pi/4 of the vector (0,1)
rotation(v = c(0, 1), theta = pi/4)
}

}
\keyword{nonparametric}
