\name{areaahulleval}
\alias{areaahulleval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Area of the alpha-convex hull
}
\description{
This function calculates the area of the \eqn{\alpha}-convex hull of a sample of points.
}
\usage{
areaahulleval(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class \code{"ahull"}.}
}
\value{
\item{area}{Area of the \eqn{\alpha}-convex hull. }
}

\seealso{\code{\link{ahull}}.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}